sms/quickstart/requirements.txt
commune-mail>=0.3.0
python-dotenv>=1.0.0
